refactor(sdk): move sequence out of entities
